% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigrtable.R
\name{bigrtable}
\alias{bigrtable}
\title{bigrtable}
\usage{
bigrtable(x, y, z)
}
\arguments{
\item{x}{The first variable of a dataset the user wants to record frequencies for}

\item{y}{The next variable of a dataset the user wants to record frequencies for}

\item{z}{The third variable of a dataset the user wants to record frequencies for}
}
\value{
A chart of the frequency values for each parameter together
}
\description{
This is a function that allows a user make a table of the frequencies of
three different variables within a dataset. Must have same number of rows.
}
\examples{
bigrtable(mtcars$mpg,mtcars$wt,mtcars$cyl)
}
