% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghist.R
\name{ghist}
\alias{ghist}
\title{ghist}
\usage{
ghist(x)
}
\arguments{
\item{x}{The variable of a dataset that's frequency values are being plotted}
}
\value{
A histogram
}
\description{
This is a function that creates a histogram of frequency values from a variable
in a dataset using ggplot2.
}
\examples{
ghist(mtcars$mpg)
}
