% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modechart.R
\name{modechart}
\alias{modechart}
\title{modechart}
\usage{
modechart(x, y, z)
}
\arguments{
\item{x}{The first variable of a dataset the user wants to record mode for}

\item{y}{The next variable of a dataset the user wants to record mode for}

\item{z}{The third variable of a dataset the user wants to record mode for}
}
\value{
A chart of the mode values for each parameter
}
\description{
This is a function that allows a user make a table of the mode values of
three different variables within a dataset.
}
\examples{
modechart(mtcars$mpg,mtcars$wt,mtcars$cyl)
}
