% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymode.R
\name{mymode}
\alias{mymode}
\title{mymode}
\usage{
mymode(x)
}
\arguments{
\item{x}{The variable of a dataset the user wants to find the mode for}
}
\value{
A value (mode)
}
\description{
This is a function that allows a user to find the mode of a variable
in a dataset.
}
\details{
Code taken from https://www.tutorialspoint.com/r/r_mean_median_mode.htm
}
\examples{
mymode(mtcars$mpg)
}
