% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myrange.R
\name{myrange}
\alias{myrange}
\title{myrange}
\usage{
myrange(x)
}
\arguments{
\item{x}{The variable of a dataset the user wants to find the range for}
}
\value{
A difference value (range)
}
\description{
This is a function that allows a user to find the range of a variable
in a dataset.
}
\examples{
myrange(mtcars$mpg)
}
